LayoutTests/imported/w3c/web-platform-tests/editing/other/white-spaces-after-execCommand-inserttext.tentative-expected.txt

a  b

PASS execCommand("inserttext", false, " ") at "a[]b"
FAIL execCommand("inserttext", false, " ") at "a []b" assert_equals: Modified text is wrong expected "a&nbsp; b" but got "a &nbsp;b"
FAIL execCommand("inserttext", false, " ") at "a&nbsp; []b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp;b" but got "a &nbsp; b"
FAIL execCommand("inserttext", false, " ") at "a&nbsp; &nbsp;[]b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; b" but got "a &nbsp; &nbsp;b"
FAIL execCommand("inserttext", false, " ") at "a&nbsp; &nbsp; []b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp;b" but got "a &nbsp; &nbsp; b"
FAIL execCommand("inserttext", false, " ") at "a&nbsp; &nbsp; &nbsp;[]b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; b" but got "a &nbsp; &nbsp; &nbsp;b"
FAIL execCommand("inserttext", false, " ") at "a&nbsp; &nbsp; &nbsp; []b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp;b" but got "a &nbsp; &nbsp; &nbsp; b"
FAIL execCommand("inserttext", false, " ") at "a&nbsp; &nbsp; &nbsp; &nbsp;[]b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp; b" but got "a &nbsp; &nbsp; &nbsp; &nbsp;b"
FAIL execCommand("inserttext", false, " ") at "a&nbsp; &nbsp; &nbsp; &nbsp; []b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp; &nbsp;b" but got "a &nbsp; &nbsp; &nbsp; &nbsp; b"
FAIL execCommand("inserttext", false, " ") at "a&nbsp; &nbsp; &nbsp; &nbsp; &nbsp;[]b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; b" but got "a &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;b"
FAIL execCommand("inserttext", false, " ") at "a&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; []b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;b" but got "a &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; b"
FAIL execCommand("inserttext", false, " ") at "a&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;[]b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; b" but got "a &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;b"
PASS execCommand("inserttext", false, " ") at "a[]"
FAIL execCommand("inserttext", false, " ") at "a&nbsp;[]" assert_equals: Modified text is wrong expected "a&nbsp;&nbsp;" but got "a &nbsp;"
FAIL execCommand("inserttext", false, " ") at "a&nbsp;&nbsp;[]" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp;" but got "a &nbsp;&nbsp;"
FAIL execCommand("inserttext", false, " ") at "a&nbsp; &nbsp;[]" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp;&nbsp;" but got "a &nbsp; &nbsp;"
FAIL execCommand("inserttext", false, " ") at "a&nbsp; &nbsp;&nbsp;[]" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp;" but got "a &nbsp; &nbsp;&nbsp;"
FAIL execCommand("inserttext", false, " ") at "a&nbsp; &nbsp; &nbsp;[]" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp;&nbsp;" but got "a &nbsp; &nbsp; &nbsp;"
FAIL execCommand("inserttext", false, " ") at "a&nbsp; &nbsp; &nbsp;&nbsp;[]" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp;" but got "a &nbsp; &nbsp; &nbsp;&nbsp;"
FAIL execCommand("inserttext", false, " ") at "a&nbsp; &nbsp; &nbsp; &nbsp;[]" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp;&nbsp;" but got "a &nbsp; &nbsp; &nbsp; &nbsp;"
FAIL execCommand("inserttext", false, " ") at "a&nbsp; &nbsp; &nbsp; &nbsp;&nbsp;[]" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp; &nbsp;" but got "a &nbsp; &nbsp; &nbsp; &nbsp;&nbsp;"
FAIL execCommand("inserttext", false, " ") at "a&nbsp; &nbsp; &nbsp; &nbsp; &nbsp;[]" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp; &nbsp;&nbsp;" but got "a &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;"
FAIL execCommand("inserttext", false, " ") at "a&nbsp; &nbsp; &nbsp; &nbsp; &nbsp;&nbsp;[]" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;" but got "a &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;&nbsp;"
FAIL execCommand("inserttext", false, " ") at "a&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;[]" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;&nbsp;" but got "a &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;"
PASS execCommand("inserttext", false, " ") at "[]b"
PASS execCommand("inserttext", false, " ") at "&nbsp;[]b"
PASS execCommand("inserttext", false, " ") at "&nbsp; []b"
PASS execCommand("inserttext", false, " ") at "&nbsp; &nbsp;[]b"
PASS execCommand("inserttext", false, " ") at "&nbsp; &nbsp; []b"
PASS execCommand("inserttext", false, " ") at "&nbsp; &nbsp; &nbsp;[]b"
PASS execCommand("inserttext", false, " ") at "&nbsp; &nbsp; &nbsp; []b"
PASS execCommand("inserttext", false, " ") at "&nbsp; &nbsp; &nbsp; &nbsp;[]b"
PASS execCommand("inserttext", false, " ") at "&nbsp; &nbsp; &nbsp; &nbsp; []b"
PASS execCommand("inserttext", false, " ") at "&nbsp; &nbsp; &nbsp; &nbsp; &nbsp;[]b"
PASS execCommand("inserttext", false, " ") at "&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; []b"
PASS execCommand("inserttext", false, " ") at "&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;[]b"
FAIL execCommand("inserttext", false, " ") at "a[] b" assert_equals: Modified text is wrong expected "a&nbsp; b" but got "a &nbsp;b"
FAIL execCommand("inserttext", false, " ") at "a&nbsp;[] b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp;b" but got "a &nbsp; b"
FAIL execCommand("inserttext", false, " ") at "a&nbsp; []&nbsp;b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; b" but got "a &nbsp; &nbsp;b"
FAIL execCommand("inserttext", false, " ") at "a&nbsp; &nbsp;[] b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp;b" but got "a &nbsp; &nbsp; b"
FAIL execCommand("inserttext", false, " ") at "a&nbsp; &nbsp; []&nbsp;b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; b" but got "a &nbsp; &nbsp; &nbsp;b"
FAIL execCommand("inserttext", false, " ") at "a&nbsp; &nbsp; &nbsp;[] b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp;b" but got "a &nbsp; &nbsp; &nbsp; b"
FAIL execCommand("inserttext", false, " ") at "a&nbsp; &nbsp; &nbsp; []&nbsp;b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp; b" but got "a &nbsp; &nbsp; &nbsp; &nbsp;b"
FAIL execCommand("inserttext", false, " ") at "a&nbsp; &nbsp; &nbsp; &nbsp;[] b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp; &nbsp;b" but got "a &nbsp; &nbsp; &nbsp; &nbsp; b"
FAIL execCommand("inserttext", false, " ") at "a&nbsp; &nbsp; &nbsp; &nbsp; []&nbsp;b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; b" but got "a &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;b"
FAIL execCommand("inserttext", false, " ") at "a&nbsp; &nbsp; &nbsp; &nbsp; &nbsp;[] b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;b" but got "a &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; b"
FAIL execCommand("inserttext", false, " ") at "a&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; []&nbsp;b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; b" but got "a &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;b"
FAIL execCommand("inserttext", false, " ") at "a&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;[] b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;b" but got "a &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; b"
FAIL execCommand("inserttext", false, " ") at "a[]&nbsp;b" assert_equals: Modified text is wrong expected "a&nbsp; b" but got "a &nbsp;b"
FAIL execCommand("inserttext", false, " ") at "a[]&nbsp;&nbsp;b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp;b" but got "a &nbsp; b"
FAIL execCommand("inserttext", false, " ") at "a&nbsp;[]&nbsp;b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp;b" but got "a &nbsp; b"
FAIL execCommand("inserttext", false, " ") at "a&nbsp;&nbsp;[]b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp;b" but got "a &nbsp; b"
FAIL execCommand("inserttext", false, " ") at "a[]&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; b" but got "a &nbsp; &nbsp; &nbsp;b"
FAIL execCommand("inserttext", false, " ") at "a&nbsp;[]&nbsp;&nbsp;&nbsp;&nbsp;b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; b" but got "a &nbsp; &nbsp; &nbsp;b"
FAIL execCommand("inserttext", false, " ") at "a&nbsp;&nbsp;[]&nbsp;&nbsp;&nbsp;b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; b" but got "a &nbsp; &nbsp; &nbsp;b"
FAIL execCommand("inserttext", false, " ") at "a&nbsp;&nbsp;&nbsp;[]&nbsp;&nbsp;b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; b" but got "a &nbsp; &nbsp; &nbsp;b"
FAIL execCommand("inserttext", false, " ") at "a&nbsp;&nbsp;&nbsp;&nbsp;[]&nbsp;b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; b" but got "a &nbsp; &nbsp; &nbsp;b"
FAIL execCommand("inserttext", false, " ") at "a&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[]b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; b" but got "a &nbsp; &nbsp; &nbsp;b"
FAIL execCommand("inserttext", false, " ") at "a[]&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp;b" but got "a &nbsp; &nbsp; &nbsp; b"
FAIL execCommand("inserttext", false, " ") at "a&nbsp;[]&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp;b" but got "a &nbsp; &nbsp; &nbsp; b"
FAIL execCommand("inserttext", false, " ") at "a&nbsp;&nbsp;[]&nbsp;&nbsp;&nbsp;&nbsp;b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp;b" but got "a &nbsp; &nbsp; &nbsp; b"
FAIL execCommand("inserttext", false, " ") at "a&nbsp;&nbsp;&nbsp;[]&nbsp;&nbsp;&nbsp;b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp;b" but got "a &nbsp; &nbsp; &nbsp; b"
FAIL execCommand("inserttext", false, " ") at "a&nbsp;&nbsp;&nbsp;&nbsp;[]&nbsp;&nbsp;b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp;b" but got "a &nbsp; &nbsp; &nbsp; b"
FAIL execCommand("inserttext", false, " ") at "a&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[]&nbsp;b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp;b" but got "a &nbsp; &nbsp; &nbsp; b"
FAIL execCommand("inserttext", false, " ") at "a&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;[]b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp;b" but got "a &nbsp; &nbsp; &nbsp; b"
PASS execCommand("inserttext", false, "\u00A0") at "a[]b"
FAIL execCommand("inserttext", false, "\u00A0") at "a&nbsp;[]b" assert_equals: Modified text is wrong expected "a&nbsp; b" but got "a &nbsp;b"
FAIL execCommand("inserttext", false, "\u00A0") at "a&nbsp; []b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp;b" but got "a &nbsp; b"
FAIL execCommand("inserttext", false, "\u00A0") at "a&nbsp; &nbsp;[]b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; b" but got "a &nbsp; &nbsp;b"
FAIL execCommand("inserttext", false, "\u00A0") at "a&nbsp; &nbsp; []b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp;b" but got "a &nbsp; &nbsp; b"
FAIL execCommand("inserttext", false, "\u00A0") at "a&nbsp; &nbsp; &nbsp;[]b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; b" but got "a &nbsp; &nbsp; &nbsp;b"
FAIL execCommand("inserttext", false, "\u00A0") at "a&nbsp; &nbsp; &nbsp; []b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp;b" but got "a &nbsp; &nbsp; &nbsp; b"
FAIL execCommand("inserttext", false, "\u00A0") at "a&nbsp; &nbsp; &nbsp; &nbsp;[]b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp; b" but got "a &nbsp; &nbsp; &nbsp; &nbsp;b"
FAIL execCommand("inserttext", false, "\u00A0") at "a&nbsp; &nbsp; &nbsp; &nbsp; []b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp; &nbsp;b" but got "a &nbsp; &nbsp; &nbsp; &nbsp; b"
FAIL execCommand("inserttext", false, "\u00A0") at "a&nbsp; &nbsp; &nbsp; &nbsp; &nbsp;[]b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; b" but got "a &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;b"
FAIL execCommand("inserttext", false, "\u00A0") at "a&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; []b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;b" but got "a &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; b"
FAIL execCommand("inserttext", false, "\u00A0") at "a&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;[]b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; b" but got "a &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;b"
PASS execCommand("inserttext", false, " ") at "a<span> []</span>b"
PASS execCommand("inserttext", false, " ") at "a<span>&nbsp; []</span>b"
PASS execCommand("inserttext", false, " ") at "a<span>&nbsp; &nbsp;[]</span>b"
PASS execCommand("inserttext", false, " ") at "a<span>&nbsp; &nbsp; []</span>b"
PASS execCommand("inserttext", false, " ") at "a<span>&nbsp; &nbsp; &nbsp;[]</span>b"
PASS execCommand("inserttext", false, " ") at "a<span>b[]</span>c"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp;[]</span>c" assert_equals: Modified text is wrong expected "a<span>b&nbsp;&nbsp;</span>c" but got "a<span>b &nbsp;</span>c"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp;&nbsp;[]</span>c" assert_equals: Modified text is wrong expected "a<span>b&nbsp; &nbsp;</span>c" but got "a<span>b &nbsp;&nbsp;</span>c"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp; &nbsp;[]</span>c" assert_equals: Modified text is wrong expected "a<span>b&nbsp; &nbsp;&nbsp;</span>c" but got "a<span>b &nbsp; &nbsp;</span>c"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp; &nbsp;&nbsp;[]</span>c" assert_equals: Modified text is wrong expected "a<span>b&nbsp; &nbsp; &nbsp;</span>c" but got "a<span>b &nbsp; &nbsp;&nbsp;</span>c"
PASS execCommand("inserttext", false, " ") at "a<span>b[]</span> c"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp;[]</span> c" assert_equals: Modified text is wrong expected "a<span>b&nbsp;&nbsp;</span> c" but got "a<span>b &nbsp;</span> c"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp;&nbsp;[]</span> c" assert_equals: Modified text is wrong expected "a<span>b&nbsp; &nbsp;</span> c" but got "a<span>b &nbsp;&nbsp;</span> c"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp; &nbsp;[]</span> c" assert_equals: Modified text is wrong expected "a<span>b&nbsp; &nbsp;&nbsp;</span> c" but got "a<span>b &nbsp; &nbsp;</span> c"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp; &nbsp;&nbsp;[]</span> c" assert_equals: Modified text is wrong expected "a<span>b&nbsp; &nbsp; &nbsp;</span> c" but got "a<span>b &nbsp; &nbsp;&nbsp;</span> c"
PASS execCommand("inserttext", false, " ") at "a<span>b[]</span>&nbsp;c"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp;[]</span>&nbsp;c" assert_equals: Modified text is wrong expected "a<span>b&nbsp;&nbsp;</span>&nbsp;c" but got "a<span>b &nbsp;</span>&nbsp;c"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp;&nbsp;[]</span>&nbsp;c" assert_equals: Modified text is wrong expected "a<span>b&nbsp; &nbsp;</span>&nbsp;c" but got "a<span>b &nbsp;&nbsp;</span>&nbsp;c"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp; &nbsp;[]</span>&nbsp;c" assert_equals: Modified text is wrong expected "a<span>b&nbsp; &nbsp;&nbsp;</span>&nbsp;c" but got "a<span>b &nbsp; &nbsp;</span>&nbsp;c"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp; &nbsp;&nbsp;[]</span>&nbsp;c" assert_equals: Modified text is wrong expected "a<span>b&nbsp; &nbsp; &nbsp;</span>&nbsp;c" but got "a<span>b &nbsp; &nbsp;&nbsp;</span>&nbsp;c"
PASS execCommand("inserttext", false, " ") at "a<span>b[]</span><span>c</span>"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp;[]</span><span>c</span>" assert_equals: Modified text is wrong expected "a<span>b&nbsp;&nbsp;</span><span>c</span>" but got "a<span>b &nbsp;</span><span>c</span>"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp;&nbsp;[]</span><span>c</span>" assert_equals: Modified text is wrong expected "a<span>b&nbsp; &nbsp;</span><span>c</span>" but got "a<span>b &nbsp;&nbsp;</span><span>c</span>"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp; &nbsp;[]</span><span>c</span>" assert_equals: Modified text is wrong expected "a<span>b&nbsp; &nbsp;&nbsp;</span><span>c</span>" but got "a<span>b &nbsp; &nbsp;</span><span>c</span>"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp; &nbsp;&nbsp;[]</span><span>c</span>" assert_equals: Modified text is wrong expected "a<span>b&nbsp; &nbsp; &nbsp;</span><span>c</span>" but got "a<span>b &nbsp; &nbsp;&nbsp;</span><span>c</span>"
PASS execCommand("inserttext", false, " ") at "a<span>b[]</span><span> c</span>"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp;[]</span><span> c</span>" assert_equals: Modified text is wrong expected "a<span>b&nbsp;&nbsp;</span><span> c</span>" but got "a<span>b &nbsp;</span><span> c</span>"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp;&nbsp;[]</span><span> c</span>" assert_equals: Modified text is wrong expected "a<span>b&nbsp; &nbsp;</span><span> c</span>" but got "a<span>b &nbsp;&nbsp;</span><span> c</span>"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp; &nbsp;[]</span><span> c</span>" assert_equals: Modified text is wrong expected "a<span>b&nbsp; &nbsp;&nbsp;</span><span> c</span>" but got "a<span>b &nbsp; &nbsp;</span><span> c</span>"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp; &nbsp;&nbsp;[]</span><span> c</span>" assert_equals: Modified text is wrong expected "a<span>b&nbsp; &nbsp; &nbsp;</span><span> c</span>" but got "a<span>b &nbsp; &nbsp;&nbsp;</span><span> c</span>"
PASS execCommand("inserttext", false, " ") at "a<span>b[]</span><span>&nbsp;c</span>"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp;[]</span><span>&nbsp;c</span>" assert_equals: Modified text is wrong expected "a<span>b&nbsp;&nbsp;</span><span>&nbsp;c</span>" but got "a<span>b &nbsp;</span><span>&nbsp;c</span>"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp;&nbsp;[]</span><span>&nbsp;c</span>" assert_equals: Modified text is wrong expected "a<span>b&nbsp; &nbsp;</span><span>&nbsp;c</span>" but got "a<span>b &nbsp;&nbsp;</span><span>&nbsp;c</span>"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp; &nbsp;[]</span><span>&nbsp;c</span>" assert_equals: Modified text is wrong expected "a<span>b&nbsp; &nbsp;&nbsp;</span><span>&nbsp;c</span>" but got "a<span>b &nbsp; &nbsp;</span><span>&nbsp;c</span>"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp; &nbsp;&nbsp;[]</span><span>&nbsp;c</span>" assert_equals: Modified text is wrong expected "a<span>b&nbsp; &nbsp; &nbsp;</span><span>&nbsp;c</span>" but got "a<span>b &nbsp; &nbsp;&nbsp;</span><span>&nbsp;c</span>"
PASS execCommand("inserttext", false, " ") at "a<span>b[]</span><span><span>c</span></span>"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp;[]</span><span><span>c</span></span>" assert_equals: Modified text is wrong expected "a<span>b&nbsp;&nbsp;</span><span><span>c</span></span>" but got "a<span>b &nbsp;</span><span><span>c</span></span>"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp;&nbsp;[]</span><span><span>c</span></span>" assert_equals: Modified text is wrong expected "a<span>b&nbsp; &nbsp;</span><span><span>c</span></span>" but got "a<span>b &nbsp;&nbsp;</span><span><span>c</span></span>"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp; &nbsp;[]</span><span><span>c</span></span>" assert_equals: Modified text is wrong expected "a<span>b&nbsp; &nbsp;&nbsp;</span><span><span>c</span></span>" but got "a<span>b &nbsp; &nbsp;</span><span><span>c</span></span>"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp; &nbsp;&nbsp;[]</span><span><span>c</span></span>" assert_equals: Modified text is wrong expected "a<span>b&nbsp; &nbsp; &nbsp;</span><span><span>c</span></span>" but got "a<span>b &nbsp; &nbsp;&nbsp;</span><span><span>c</span></span>"
PASS execCommand("inserttext", false, " ") at "a<span>b[]</span><span><span> c</span></span>"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp;[]</span><span><span> c</span></span>" assert_equals: Modified text is wrong expected "a<span>b&nbsp;&nbsp;</span><span><span> c</span></span>" but got "a<span>b &nbsp;</span><span><span> c</span></span>"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp;&nbsp;[]</span><span><span> c</span></span>" assert_equals: Modified text is wrong expected "a<span>b&nbsp; &nbsp;</span><span><span> c</span></span>" but got "a<span>b &nbsp;&nbsp;</span><span><span> c</span></span>"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp; &nbsp;[]</span><span><span> c</span></span>" assert_equals: Modified text is wrong expected "a<span>b&nbsp; &nbsp;&nbsp;</span><span><span> c</span></span>" but got "a<span>b &nbsp; &nbsp;</span><span><span> c</span></span>"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp; &nbsp;&nbsp;[]</span><span><span> c</span></span>" assert_equals: Modified text is wrong expected "a<span>b&nbsp; &nbsp; &nbsp;</span><span><span> c</span></span>" but got "a<span>b &nbsp; &nbsp;&nbsp;</span><span><span> c</span></span>"
PASS execCommand("inserttext", false, " ") at "a<span>b[]</span><span><span>&nbsp;c</span></span>"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp;[]</span><span><span>&nbsp;c</span></span>" assert_equals: Modified text is wrong expected "a<span>b&nbsp;&nbsp;</span><span><span>&nbsp;c</span></span>" but got "a<span>b &nbsp;</span><span><span>&nbsp;c</span></span>"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp;&nbsp;[]</span><span><span>&nbsp;c</span></span>" assert_equals: Modified text is wrong expected "a<span>b&nbsp; &nbsp;</span><span><span>&nbsp;c</span></span>" but got "a<span>b &nbsp;&nbsp;</span><span><span>&nbsp;c</span></span>"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp; &nbsp;[]</span><span><span>&nbsp;c</span></span>" assert_equals: Modified text is wrong expected "a<span>b&nbsp; &nbsp;&nbsp;</span><span><span>&nbsp;c</span></span>" but got "a<span>b &nbsp; &nbsp;</span><span><span>&nbsp;c</span></span>"
FAIL execCommand("inserttext", false, " ") at "a<span>b&nbsp; &nbsp;&nbsp;[]</span><span><span>&nbsp;c</span></span>" assert_equals: Modified text is wrong expected "a<span>b&nbsp; &nbsp; &nbsp;</span><span><span>&nbsp;c</span></span>" but got "a<span>b &nbsp; &nbsp;&nbsp;</span><span><span>&nbsp;c</span></span>"
PASS execCommand("inserttext", false, " ") at "a<span><span>b[]</span></span><span>c</span>"
FAIL execCommand("inserttext", false, " ") at "a<span><span>b&nbsp;[]</span></span><span>c</span>" assert_equals: Modified text is wrong expected "a<span><span>b&nbsp;&nbsp;</span></span><span>c</span>" but got "a<span><span>b &nbsp;</span></span><span>c</span>"
FAIL execCommand("inserttext", false, " ") at "a<span><span>b&nbsp;&nbsp;[]</span></span><span>c</span>" assert_equals: Modified text is wrong expected "a<span><span>b&nbsp; &nbsp;</span></span><span>c</span>" but got "a<span><span>b &nbsp;&nbsp;</span></span><span>c</span>"
FAIL execCommand("inserttext", false, " ") at "a<span><span>b&nbsp; &nbsp;[]</span></span><span>c</span>" assert_equals: Modified text is wrong expected "a<span><span>b&nbsp; &nbsp;&nbsp;</span></span><span>c</span>" but got "a<span><span>b &nbsp; &nbsp;</span></span><span>c</span>"
FAIL execCommand("inserttext", false, " ") at "a<span><span>b&nbsp; &nbsp;&nbsp;[]</span></span><span>c</span>" assert_equals: Modified text is wrong expected "a<span><span>b&nbsp; &nbsp; &nbsp;</span></span><span>c</span>" but got "a<span><span>b &nbsp; &nbsp;&nbsp;</span></span><span>c</span>"
PASS execCommand("inserttext", false, " ") at "a<span><span>b[]</span></span><span> c</span>"
FAIL execCommand("inserttext", false, " ") at "a<span><span>b&nbsp;[]</span></span><span> c</span>" assert_equals: Modified text is wrong expected "a<span><span>b&nbsp;&nbsp;</span></span><span> c</span>" but got "a<span><span>b &nbsp;</span></span><span> c</span>"
FAIL execCommand("inserttext", false, " ") at "a<span><span>b&nbsp;&nbsp;[]</span></span><span> c</span>" assert_equals: Modified text is wrong expected "a<span><span>b&nbsp; &nbsp;</span></span><span> c</span>" but got "a<span><span>b &nbsp;&nbsp;</span></span><span> c</span>"
FAIL execCommand("inserttext", false, " ") at "a<span><span>b&nbsp; &nbsp;[]</span></span><span> c</span>" assert_equals: Modified text is wrong expected "a<span><span>b&nbsp; &nbsp;&nbsp;</span></span><span> c</span>" but got "a<span><span>b &nbsp; &nbsp;</span></span><span> c</span>"
FAIL execCommand("inserttext", false, " ") at "a<span><span>b&nbsp; &nbsp;&nbsp;[]</span></span><span> c</span>" assert_equals: Modified text is wrong expected "a<span><span>b&nbsp; &nbsp; &nbsp;</span></span><span> c</span>" but got "a<span><span>b &nbsp; &nbsp;&nbsp;</span></span><span> c</span>"
PASS execCommand("inserttext", false, " ") at "a<span><span>b[]</span></span><span>&nbsp;c</span>"
FAIL execCommand("inserttext", false, " ") at "a<span><span>b&nbsp;[]</span></span><span>&nbsp;c</span>" assert_equals: Modified text is wrong expected "a<span><span>b&nbsp;&nbsp;</span></span><span>&nbsp;c</span>" but got "a<span><span>b &nbsp;</span></span><span>&nbsp;c</span>"
FAIL execCommand("inserttext", false, " ") at "a<span><span>b&nbsp;&nbsp;[]</span></span><span>&nbsp;c</span>" assert_equals: Modified text is wrong expected "a<span><span>b&nbsp; &nbsp;</span></span><span>&nbsp;c</span>" but got "a<span><span>b &nbsp;&nbsp;</span></span><span>&nbsp;c</span>"
FAIL execCommand("inserttext", false, " ") at "a<span><span>b&nbsp; &nbsp;[]</span></span><span>&nbsp;c</span>" assert_equals: Modified text is wrong expected "a<span><span>b&nbsp; &nbsp;&nbsp;</span></span><span>&nbsp;c</span>" but got "a<span><span>b &nbsp; &nbsp;</span></span><span>&nbsp;c</span>"
FAIL execCommand("inserttext", false, " ") at "a<span><span>b&nbsp; &nbsp;&nbsp;[]</span></span><span>&nbsp;c</span>" assert_equals: Modified text is wrong expected "a<span><span>b&nbsp; &nbsp; &nbsp;</span></span><span>&nbsp;c</span>" but got "a<span><span>b &nbsp; &nbsp;&nbsp;</span></span><span>&nbsp;c</span>"
FAIL execCommand("inserttext", false, "  ") at "a[]b" assert_equals: Modified text is wrong expected "a&nbsp; b" but got "a &nbsp;b"
FAIL execCommand("inserttext", false, "   ") at "a[]b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp;b" but got "a &nbsp; b"
FAIL execCommand("inserttext", false, "    ") at "a[]b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; b" but got "a &nbsp; &nbsp;b"
FAIL execCommand("inserttext", false, "     ") at "a[]b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp;b" but got "a &nbsp; &nbsp; b"
FAIL execCommand("inserttext", false, "      ") at "a[]b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; b" but got "a &nbsp; &nbsp; &nbsp;b"
FAIL execCommand("inserttext", false, "       ") at "a[]b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp;b" but got "a &nbsp; &nbsp; &nbsp; b"
FAIL execCommand("inserttext", false, "  ") at "a[]" assert_equals: Modified text is wrong expected "a&nbsp;&nbsp;" but got "a &nbsp;"
FAIL execCommand("inserttext", false, "   ") at "a[]" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp;" but got "a &nbsp;&nbsp;"
FAIL execCommand("inserttext", false, "    ") at "a[]" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp;&nbsp;" but got "a &nbsp; &nbsp;"
FAIL execCommand("inserttext", false, "     ") at "a[]" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp;" but got "a &nbsp; &nbsp;&nbsp;"
FAIL execCommand("inserttext", false, "      ") at "a[]" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp;&nbsp;" but got "a &nbsp; &nbsp; &nbsp;"
FAIL execCommand("inserttext", false, "       ") at "a[]" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp;" but got "a &nbsp; &nbsp; &nbsp;&nbsp;"
FAIL execCommand("inserttext", false, "\u00A0\u00A0") at "a[]b" assert_equals: Modified text is wrong expected "a&nbsp; b" but got "a &nbsp;b"
FAIL execCommand("inserttext", false, "\u00A0\u00A0\u00A0") at "a[]b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp;b" but got "a &nbsp; b"
FAIL execCommand("inserttext", false, "\u00A0\u00A0\u00A0\u00A0") at "a[]b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; b" but got "a &nbsp; &nbsp;b"
FAIL execCommand("inserttext", false, "\u00A0\u00A0\u00A0\u00A0\u00A0") at "a[]b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp;b" but got "a &nbsp; &nbsp; b"
FAIL execCommand("inserttext", false, "\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0") at "a[]b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; b" but got "a &nbsp; &nbsp; &nbsp;b"
FAIL execCommand("inserttext", false, "\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0") at "a[]b" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp;b" but got "a &nbsp; &nbsp; &nbsp; b"
FAIL execCommand("inserttext", false, "\u00A0\u00A0") at "a[]" assert_equals: Modified text is wrong expected "a&nbsp;&nbsp;" but got "a &nbsp;"
FAIL execCommand("inserttext", false, "\u00A0\u00A0\u00A0") at "a[]" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp;" but got "a &nbsp;&nbsp;"
FAIL execCommand("inserttext", false, "\u00A0\u00A0\u00A0\u00A0") at "a[]" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp;&nbsp;" but got "a &nbsp; &nbsp;"
FAIL execCommand("inserttext", false, "\u00A0\u00A0\u00A0\u00A0\u00A0") at "a[]" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp;" but got "a &nbsp; &nbsp;&nbsp;"
FAIL execCommand("inserttext", false, "\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0") at "a[]" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp;&nbsp;" but got "a &nbsp; &nbsp; &nbsp;"
FAIL execCommand("inserttext", false, "\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0") at "a[]" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp; &nbsp;" but got "a &nbsp; &nbsp; &nbsp;&nbsp;"
PASS execCommand("inserttext", false, " ") at "a[]<span style=white-space:pre>b</span>"
FAIL execCommand("inserttext", false, " ") at "a&nbsp;[]<span style=white-space:pre>b</span>" assert_equals: Modified text is wrong expected "a&nbsp;&nbsp;<span style=\"white-space:pre\">b</span>" but got "a &nbsp;<span style=\"white-space:pre\">b</span>"
FAIL execCommand("inserttext", false, " ") at "a&nbsp;&nbsp;[]<span style=white-space:pre>b</span>" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp;<span style=\"white-space:pre\">b</span>" but got "a &nbsp;&nbsp;<span style=\"white-space:pre\">b</span>"
FAIL execCommand("inserttext", false, " ") at "a&nbsp; &nbsp;[]<span style=white-space:pre>b</span>" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp;&nbsp;<span style=\"white-space:pre\">b</span>" but got "a &nbsp; &nbsp;<span style=\"white-space:pre\">b</span>"
FAIL execCommand("inserttext", false, " ") at "a&nbsp; &nbsp;&nbsp;[]<span style=white-space:pre>b</span>" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp;<span style=\"white-space:pre\">b</span>" but got "a &nbsp; &nbsp;&nbsp;<span style=\"white-space:pre\">b</span>"
PASS execCommand("inserttext", false, " ") at "a[]<span style=white-space:pre> </span>"
FAIL execCommand("inserttext", false, " ") at "a&nbsp;[]<span style=white-space:pre> </span>" assert_equals: Modified text is wrong expected "a&nbsp;&nbsp;<span style=\"white-space:pre\"> </span>" but got "a &nbsp;<span style=\"white-space:pre\"> </span>"
FAIL execCommand("inserttext", false, " ") at "a&nbsp;&nbsp;[]<span style=white-space:pre> </span>" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp;<span style=\"white-space:pre\"> </span>" but got "a &nbsp;&nbsp;<span style=\"white-space:pre\"> </span>"
FAIL execCommand("inserttext", false, " ") at "a&nbsp; &nbsp;[]<span style=white-space:pre> </span>" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp;&nbsp;<span style=\"white-space:pre\"> </span>" but got "a &nbsp; &nbsp;<span style=\"white-space:pre\"> </span>"
FAIL execCommand("inserttext", false, " ") at "a&nbsp; &nbsp;&nbsp;[]<span style=white-space:pre> </span>" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp;<span style=\"white-space:pre\"> </span>" but got "a &nbsp; &nbsp;&nbsp;<span style=\"white-space:pre\"> </span>"
PASS execCommand("inserttext", false, " ") at "a[]<span style=white-space:pre>&nbsp;</span>"
FAIL execCommand("inserttext", false, " ") at "a&nbsp;[]<span style=white-space:pre>&nbsp;</span>" assert_equals: Modified text is wrong expected "a&nbsp;&nbsp;<span style=\"white-space:pre\">&nbsp;</span>" but got "a &nbsp;<span style=\"white-space:pre\">&nbsp;</span>"
FAIL execCommand("inserttext", false, " ") at "a&nbsp;&nbsp;[]<span style=white-space:pre>&nbsp;</span>" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp;<span style=\"white-space:pre\">&nbsp;</span>" but got "a &nbsp;&nbsp;<span style=\"white-space:pre\">&nbsp;</span>"
FAIL execCommand("inserttext", false, " ") at "a&nbsp; &nbsp;[]<span style=white-space:pre>&nbsp;</span>" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp;&nbsp;<span style=\"white-space:pre\">&nbsp;</span>" but got "a &nbsp; &nbsp;<span style=\"white-space:pre\">&nbsp;</span>"
FAIL execCommand("inserttext", false, " ") at "a&nbsp; &nbsp;&nbsp;[]<span style=white-space:pre>&nbsp;</span>" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp;<span style=\"white-space:pre\">&nbsp;</span>" but got "a &nbsp; &nbsp;&nbsp;<span style=\"white-space:pre\">&nbsp;</span>"
PASS execCommand("inserttext", false, "\u00A0") at "a[]<span style=white-space:pre>b</span>"
FAIL execCommand("inserttext", false, "\u00A0") at "a&nbsp;[]<span style=white-space:pre>b</span>" assert_equals: Modified text is wrong expected "a&nbsp;&nbsp;<span style=\"white-space:pre\">b</span>" but got "a &nbsp;<span style=\"white-space:pre\">b</span>"
FAIL execCommand("inserttext", false, "\u00A0") at "a&nbsp;&nbsp;[]<span style=white-space:pre>b</span>" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp;<span style=\"white-space:pre\">b</span>" but got "a &nbsp;&nbsp;<span style=\"white-space:pre\">b</span>"
FAIL execCommand("inserttext", false, "\u00A0") at "a&nbsp; &nbsp;[]<span style=white-space:pre>b</span>" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp;&nbsp;<span style=\"white-space:pre\">b</span>" but got "a &nbsp; &nbsp;<span style=\"white-space:pre\">b</span>"
FAIL execCommand("inserttext", false, "\u00A0") at "a&nbsp; &nbsp;&nbsp;[]<span style=white-space:pre>b</span>" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp;<span style=\"white-space:pre\">b</span>" but got "a &nbsp; &nbsp;&nbsp;<span style=\"white-space:pre\">b</span>"
PASS execCommand("inserttext", false, "\u00A0") at "a[]<span style=white-space:pre> </span>"
FAIL execCommand("inserttext", false, "\u00A0") at "a&nbsp;[]<span style=white-space:pre> </span>" assert_equals: Modified text is wrong expected "a&nbsp;&nbsp;<span style=\"white-space:pre\"> </span>" but got "a &nbsp;<span style=\"white-space:pre\"> </span>"
FAIL execCommand("inserttext", false, "\u00A0") at "a&nbsp;&nbsp;[]<span style=white-space:pre> </span>" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp;<span style=\"white-space:pre\"> </span>" but got "a &nbsp;&nbsp;<span style=\"white-space:pre\"> </span>"
FAIL execCommand("inserttext", false, "\u00A0") at "a&nbsp; &nbsp;[]<span style=white-space:pre> </span>" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp;&nbsp;<span style=\"white-space:pre\"> </span>" but got "a &nbsp; &nbsp;<span style=\"white-space:pre\"> </span>"
FAIL execCommand("inserttext", false, "\u00A0") at "a&nbsp; &nbsp;&nbsp;[]<span style=white-space:pre> </span>" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp;<span style=\"white-space:pre\"> </span>" but got "a &nbsp; &nbsp;&nbsp;<span style=\"white-space:pre\"> </span>"
PASS execCommand("inserttext", false, "\u00A0") at "a[]<span style=white-space:pre>&nbsp;</span>"
FAIL execCommand("inserttext", false, "\u00A0") at "a&nbsp;[]<span style=white-space:pre>&nbsp;</span>" assert_equals: Modified text is wrong expected "a&nbsp;&nbsp;<span style=\"white-space:pre\">&nbsp;</span>" but got "a &nbsp;<span style=\"white-space:pre\">&nbsp;</span>"
FAIL execCommand("inserttext", false, "\u00A0") at "a&nbsp;&nbsp;[]<span style=white-space:pre>&nbsp;</span>" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp;<span style=\"white-space:pre\">&nbsp;</span>" but got "a &nbsp;&nbsp;<span style=\"white-space:pre\">&nbsp;</span>"
FAIL execCommand("inserttext", false, "\u00A0") at "a&nbsp; &nbsp;[]<span style=white-space:pre>&nbsp;</span>" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp;&nbsp;<span style=\"white-space:pre\">&nbsp;</span>" but got "a &nbsp; &nbsp;<span style=\"white-space:pre\">&nbsp;</span>"
FAIL execCommand("inserttext", false, "\u00A0") at "a&nbsp; &nbsp;&nbsp;[]<span style=white-space:pre>&nbsp;</span>" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; &nbsp;<span style=\"white-space:pre\">&nbsp;</span>" but got "a &nbsp; &nbsp;&nbsp;<span style=\"white-space:pre\">&nbsp;</span>"
PASS execCommand("inserttext", false, "b") at "a&nbsp;[]&nbsp;c"
FAIL execCommand("inserttext", false, "b") at "a[]&nbsp;&nbsp;&nbsp;&nbsp;c" assert_equals: Modified text is wrong expected "ab&nbsp; &nbsp; c" but got "ab &nbsp; &nbsp;c"
FAIL execCommand("inserttext", false, "b") at "a&nbsp;[]&nbsp;&nbsp;&nbsp;c" assert_equals: Modified text is wrong expected "a b&nbsp; &nbsp;c" but got "a b &nbsp; c"
FAIL execCommand("inserttext", false, "b") at "a&nbsp;&nbsp;[]&nbsp;&nbsp;c" assert_equals: Modified text is wrong expected "a&nbsp; b&nbsp; c" but got "a &nbsp;b &nbsp;c"
FAIL execCommand("inserttext", false, "b") at "a&nbsp;&nbsp;&nbsp;[]&nbsp;c" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp;b c" but got "a &nbsp; b c"
FAIL execCommand("inserttext", false, "b") at "a&nbsp;&nbsp;&nbsp;&nbsp;[]c" assert_equals: Modified text is wrong expected "a&nbsp; &nbsp; bc" but got "a &nbsp; &nbsp;bc"
PASS execCommand("inserttext", false, ""): "a[]|b"
PASS execCommand("inserttext", false, ""): "a []|b"
PASS execCommand("inserttext", false, ""): "a |[]b"
FAIL execCommand("inserttext", false, ""): "a[]&nbsp;|b" assert_equals: Modified text is wrong expected "a&nbsp;" but got "a "
FAIL execCommand("inserttext", false, ""): "a&nbsp;[]|b" assert_equals: Modified text is wrong expected "a&nbsp;" but got "a "
FAIL execCommand("inserttext", false, ""): "a&nbsp;|[]b" assert_equals: Modified text is wrong expected "a&nbsp;" but got "a "
PASS execCommand("inserttext", false, ""): "a[]|&nbsp;b"
PASS execCommand("inserttext", false, ""): "a|[]&nbsp;b"
PASS execCommand("inserttext", false, ""): "a|&nbsp;[]b"
FAIL execCommand("inserttext", false, ""): "a[] |&nbsp;b" assert_equals: Modified text is wrong expected "a " but got "a&nbsp;"
FAIL execCommand("inserttext", false, ""): "a []|&nbsp;b" assert_equals: Modified text is wrong expected "a " but got "a&nbsp;"
FAIL execCommand("inserttext", false, ""): "a |[]&nbsp;b" assert_equals: Modified text is wrong expected "a " but got "a&nbsp;"
PASS execCommand("inserttext", false, ""): "a |&nbsp;[]b"
PASS execCommand("inserttext", false, ""): "a[]&nbsp;| b"
PASS execCommand("inserttext", false, ""): "a&nbsp;[]| b"
PASS execCommand("inserttext", false, ""): "a&nbsp;|[] b"
FAIL execCommand("inserttext", false, ""): "a&nbsp;| []b" assert_equals: Modified text is wrong expected " b" but got "&nbsp;b"
PASS execCommand("inserttext", false, ""): "a[]&nbsp;|&nbsp;b"
PASS execCommand("inserttext", false, ""): "a&nbsp;[]|&nbsp;b"
PASS execCommand("inserttext", false, ""): "a&nbsp;|[]&nbsp;b"
PASS execCommand("inserttext", false, ""): "a&nbsp;|&nbsp;[]b"
